Added Eclipse cleanup profile
And cleaned up all the old code, using said profile

+++ MODSRC/vazkii/tinkerer/client/particle/EntityFXColoredPortal.java
@@ -8,10 +8,9 @@ package vazkii.tinkerer.client.particle;
 
 import java.awt.Color;
 
-import vazkii.tinkerer.helper.MiscHelper;
-
 import net.minecraft.client.particle.EntityPortalFX;
 import net.minecraft.world.World;
+import vazkii.tinkerer.helper.MiscHelper;
 
 /**
  * EntityFXColoredPortal
@@ -26,12 +25,12 @@ public class EntityFXColoredPortal extends EntityPortalFX {
 		EntityFXColoredPortal entity = new EntityFXColoredPortal(color, world, x, y, z, motionX, motionY, motionZ);
 		MiscHelper.getMc().effectRenderer.addEffect(entity);
 	}
-	
+
 	public EntityFXColoredPortal(Color color, World par1World, double par2, double par4, double par6, double par8, double par10, double par12) {
 		super(par1World, par2, par4, par6, par8, par10, par12);
-		particleRed = (float)color.getRed() / 255F; // Particle colors are 0F-1F, RGB colors are 0-255
-		particleGreen =(float)color.getGreen() / 255F;
-		particleBlue =  (float)color.getBlue() / 255F;
+		particleRed = color.getRed() / 255F; // Particle colors are 0F-1F, RGB colors are 0-255
+		particleGreen =color.getGreen() / 255F;
+		particleBlue =  color.getBlue() / 255F;
 	}
 
 }
